refactor: share parse_literal test helper
