+++ crates/cli/tests/block_size.rs
@@ -1,17 +1,9 @@
 // crates/cli/tests/block_size.rs
+
 use assert_cmd::Command;
 use tempfile::tempdir;
-
-fn parse_literal(stats: &str) -> usize {
-    for line in stats.lines() {
-        let line = line.trim();
-        if let Some(rest) = line.strip_prefix("Literal data: ") {
-            let num_str = rest.split_whitespace().next().unwrap().replace(",", "");
-            return num_str.parse().unwrap();
-        }
-    }
-    panic!("no literal data in stats: {stats}");
-}
+mod common;
+use common::parse_literal;
 
 #[test]
 fn block_size_literal_data_matches() {
