+++ tests/block_size.rs
@@ -8,20 +8,8 @@ use engine::{Op, SyncOptions, block_size, compute_delta, sync};
 use filters::Matcher;
 use std::fs;
 use tempfile::tempdir;
-
-fn parse_literal(stats: &str) -> usize {
-    for line in stats.lines() {
-        let line = line.trim();
-        if let Some(rest) = line
-            .strip_prefix("Literal data: ")
-            .or_else(|| line.strip_prefix("Unmatched data: "))
-        {
-            let num_str = rest.split_whitespace().next().unwrap().replace(",", "");
-            return num_str.parse().unwrap();
-        }
-    }
-    panic!("no literal data in stats: {stats}");
-}
+mod common;
+use common::parse_literal;
 
 #[test]
 fn block_size_matches_upstream() {
